mic_array_dashboard/backend/requirements.txt
tornado>=6.3
pyserial>=3.5
sounddevice>=0.4
